requirements.txt
discord-webhook==0.5.0
PySocks==1.6.8
PyYAML==5.1
requests==2.21.0
requests-toolbelt==0.9.1